requirements.txt
Flask
Gunicorn
discord.py